some improvements

+++ Implements/Characters/DefaultCharacterHealth.cs
@@ -5,17 +5,17 @@ namespace Hathor
 {
     class DefaultCharacterHealthBuffer : ICharacterHealthBuffer
     {
-        protected int mValue;
+        protected float mValue;
 
         public DefaultCharacterHealthBuffer()
         {
             this.mValue = 0;
         }
 
-        public int Value { get => this.mValue; }
+        public float Value { get => this.mValue; }
 
         // 提升数值
-        public void Increase(int value)
+        public void Increase(float value)
         {
             this.mValue += value;
         }
@@ -23,7 +23,7 @@ namespace Hathor
         // 扩大提升数值
         public void Expend(float value)
         {
-            this.mValue = (int)((float)(this.mValue) * value);
+            this.mValue = this.mValue * value;
         }
     }
 
@@ -31,14 +31,14 @@ namespace Hathor
     {
         protected ICharacter mChar;
         protected IAttribute mMaxHP;
-        protected int mHP;
+        protected float mHP;
         protected Dictionary<string, ICharacterHealthBuffer> mHealingBuffers;
         protected Dictionary<string, ICharacterHealthBuffer> mDamageBuffers;
 
         public DefaultCharacterHealth(
             ICharacter character,
             IAttribute maxHp,
-            int hp
+            float hp
         )
         {
             this.mChar = character;
@@ -48,17 +48,17 @@ namespace Hathor
             this.mDamageBuffers = new Dictionary<string, ICharacterHealthBuffer>();
         }
 
-        protected int FlushHealingBuffer()
+        protected float FlushHealingBuffer()
         {
             if (this.mHealingBuffers.Count == 0)
             {
                 return 0;
             }
 
-            int value = 0;
+            float value = 0;
             foreach (var heal in this.mHealingBuffers)
             {
-                int healValue = heal.Value.Value;
+                float healValue = heal.Value.Value;
                 if (healValue <= 0)
                 {
                     continue;
@@ -70,17 +70,17 @@ namespace Hathor
             return value;
         }
 
-        protected int FlushDamageBuffer()
+        protected float FlushDamageBuffer()
         {
             if (this.mDamageBuffers.Count == 0)
             {
                 return 0;
             }
 
-            int value = 0;
+            float value = 0;
             foreach (var damage in this.mDamageBuffers)
             {
-                int damageValue = damage.Value.Value;
+                float damageValue = damage.Value.Value;
                 if (damageValue <= 0)
                 {
                     continue;
@@ -97,10 +97,10 @@ namespace Hathor
             return value;
         }
 
-        protected int GetValue()
+        protected float GetValue()
         {
             // 更新血量 -------------------------------------
-            int HPDelta = this.FlushHealingBuffer() - this.FlushDamageBuffer();
+            float HPDelta = this.FlushHealingBuffer() - this.FlushDamageBuffer();
             if (HPDelta != 0) // 没有变化就不用更新了
             {
                 if (this.mHP >= this.mMaxHP.Value)
@@ -118,7 +118,7 @@ namespace Hathor
         }
 
         // 血量值
-        public int Value { get => this.GetValue(); }
+        public float Value { get => this.GetValue(); }
 
         // 回复计算
         public ICharacterHealthBuffer GetHealingBuffer(string series)
