+++ Implements/Events/BattleEvent.cs
@@ -3,7 +3,7 @@ namespace Hathor
     class BattleDamageEvent : IEvent
     {
         public ICharacter Character = null;
-        public int Damage = 0;
+        public float Damage = 0;
         public string Series { get => "battle"; }
 
         public override string ToString()
